Final fix to SortOrder problems: these occurred because sort order was saved in a settings object which is not available before persistance, and thus not available outside the web-app.  (Bugzid: 400)
It is not trivially possible to split SortOrder's into those that are system-defined, and those that are user defined: so solution is to remove persistance of sortorder.  SortOrder's should be split just like Filter's should be split (see Bugzid:396).

Finally, this means persistance of GenericLijstManagerBase is incomplete (Bugzid:344), but cannot improve before solving Case 396.