- Rewrote Symmetric.cs; those 700 lines essentially come down to: return string.Join("", 	new MD5CryptoServiceProvider().ComputeHash(Encoding.UTF8.GetBytes(plainText)) 		.Select(byteVal=>byteVal.ToString("x2")) 		.ToArray());
Replaced instances of SqlCommand in data reading api with a custom class DataCommand.  This class can now track execution time of any queries on a page.  DataConnectie, BusinessConnection and SessionManager adapted to pass this information forward to the user interface - ONLY in debug builds, to avoid any possible security implications.  Parameters are not printed.

- removed some unused code.

- refactored authentication using MD5's to all refer to Account in the same fashion.

- bugfix for ToolTips.js; when a tooltip is taller than the screen, now overflow at the bottom, not the top (you can scroll there, after all).