SimplifyMetaDataKoppelTabelLookup
avoid special casing NxN koppeltabellen - the only difference was that
those don't have a default empty item with id nil; and by listing those
in Applicatie.KoppelTabel instead of Applicatie.KoppelTabelTabel we
avoid the empty default item and (as a bonus) replace two datatables
with IMetaObjects.

This is a precursor cleanup to fixing nightlies studielinkstatus errors.