Define interface IBuildableQuery
This is a step towards a less fancy QueryBuilder.  Rather than
specifying that each query must be roughly analogous to a singly linked
list of query parts and then specify that query parts are sql or
parameters, instead specify the "bare minimum", that a query needs to be
able to append itself to an sql accumulator that is little more than an
append-only dictionary of params and a stringbuilder.