Get attributes via enumMetaCache
This is faster, but more importantly, it's consistent with how we get
attributes elsewhere.

The API change that causes most changes is that `GetAttributeValue` is
replaced by methods that do no return a single attribute, but instead
return all attributes (or all attibutes of a given type).  This is more
correct, since it is possible to define an attribute multiple times on a
single target.

@tjibbechris @markijbema