Cache delegate
the delegate mapping a column ordering to DbDataReader unpacker is
simply a wrapper around a method.  Due to arcane implementation details
of .NET internals, method group conversion to a delegate can be
expensive if you aren't on one of the special-case fast paths.  We are
not here, and that means that the trivial method-to-delegate conversion
ends up taking a ridiculous 4% of time for single-row result set
queries.

Solution: create the delegate wrapper around the method just once.