Use custom string equality
For some reason, StringComparer.OrdinalIgnoreCase is relatively
expensive, and shows up in profiles when it comes to selecting a
DbDataReader unpacker for a given column ordering (we need to check
equality every query for every column).

It's possible the .net algorithm has a high startup cost and all our
strings are short; or that it goes to great lengths to generate
high-quality hashes, but neither is useful to us here: almost all query
types will have very few column orderings, and so a "decent" hashcode
implementation that's fast is enough.  This trivial implementation is
based on java's String.hashCode(), except that ascii case insensitivity
is ensured.

Similarly, by the time the string equality is checked, the hashes must
match, so our string equality is almost always called with identical
strings, or string that differ only in few places and there only in
casing (i.e. not semantically different).

It's a little weird that trivial implementations should beat the built
in methods, but this alone improves query perf by a few percent.