Allow safe copy-&-mutate operation
unfortunately the struct+mutation solution is a little too simple.  The idea *was* that it was OK to have a mutable struct - and it is as long as it's used by value.  But because the mutation can occur through an interface, and because that means it can occur on *boxed* values aka references, you can get pretty confusing aliasing with mutaiton.