Remove legacy header
I can't get this to work on FF; and it's working fine on chrome without the legacy header.  MDN doesn't explain why; but caniuse claims safari and FF only support certain allow statements, which is pointless.

Since the header is deprecated anyhow; let's just stick with the standardized solution, even if it's chrome-only for now.